requirements.txt
pygame==2.5.1
pyserial==3.5